+++ src/Stormancer.Plugins/GameFinder/Stormancer.Server.Plugins.GameFinder/Interfaces/IGameFinderService.cs
@@ -54,8 +54,7 @@ namespace Stormancer.Server.Plugins.GameFinder
         /// <param name="party"></param>
         /// <param name="ct"></param>
         /// <returns></returns>
-        Task FindGame(Party party, CancellationToken ct);
-
+        Task<FindGameResult> FindGame(Party party, CancellationToken ct);
       
         /// <summary>
         /// Cancels a pending game search.