+++ src/Stormancer.Plugins/Profile/Stormancer.Server.Plugins.Profile/UserProfilePart.cs
@@ -22,6 +22,7 @@
 
 using Newtonsoft.Json.Linq;
 using Stormancer.Server.Plugins.Users;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -30,12 +31,10 @@ namespace Stormancer.Server.Plugins.Profile
     class PseudoProfilePart : IProfilePartBuilder
     {
         private readonly IUserService _users;
-        private readonly IUserSessions _sessions;
 
-        public PseudoProfilePart(IUserService users, IUserSessions sessions)
+        public PseudoProfilePart(IUserService users)
         {
             _users = users;
-            _sessions = sessions;
         }
 
         public async Task GetProfiles(ProfileCtx ctx, CancellationToken cancellationToken)
@@ -45,35 +44,37 @@ namespace Stormancer.Server.Plugins.Profile
                 return;
             }
 
-            //TODO: Optimize.
-            foreach (var id in ctx.Users)
+            var allUsers = await _users.GetUsers(ctx.Users, cancellationToken);
+            var users = allUsers.Where(kvp => kvp.Value != null);
+
+            foreach (var pair in users)
             {
-                // Prefer to retrieve the user directly from the session. If the user is offline, use the database.
-                var session = await _sessions.GetSessionByUserId(id, cancellationToken);
-                var user = session?.User ?? await _users.GetUser(id);
-                ctx.UpdateProfileData(id, "user", j =>
-                  {
-                      if (user != null)
-                      {
-                          j["lastPlatform"] = user.LastPlatform ?? "";
+                var userId = pair.Key;
+                var user = pair.Value;
 
-                          if (!j.ContainsKey("platforms"))
-                          {
-                              j["platforms"] = new JObject();
-                          }
+                ctx.UpdateProfileData(userId, "user", j =>
+                {
+                    if (user != null)
+                    {
+                        j["lastPlatform"] = user.LastPlatform ?? "";
 
-                          if (!j.ContainsKey("handle") && user.UserData.ContainsKey("handle"))
-                          {
-                              j["userhandle"] = user.UserData["handle"];
-                          }
+                        if (!j.ContainsKey("platforms"))
+                        {
+                            j["platforms"] = new JObject();
+                        }
 
-                          if (!j.ContainsKey("pseudo") && user.UserData.ContainsKey("pseudo"))
-                          {
-                              j["pseudo"] = user.UserData["pseudo"];
-                          }
-                      }
-                      return j;
-                  });
+                        if (!j.ContainsKey("handle") && user.UserData.ContainsKey("handle"))
+                        {
+                            j["userhandle"] = user.UserData["handle"];
+                        }
+
+                        if (!j.ContainsKey("pseudo") && user.UserData.ContainsKey("pseudo"))
+                        {
+                            j["pseudo"] = user.UserData["pseudo"];
+                        }
+                    }
+                    return j;
+                });
             }
         }
     }