+++ src/Stormancer.Plugins/GameFinder/Stormancer.Server.Plugins.GameFinder/Changelog.rst
@@ -9,6 +9,16 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+7.3.0
+-----
+Changed
+*******
+- FindGame returns false and an error message on failure
+Added
+*****
+- Added LeaderPartyIsHost parameter to party gamefinder.
+- Added "gamefinder" "end" analytics event.
+
 7.2.1.1
 -------
 Fixed