+++ src/Stormancer.Plugins/Leaderboards/Stormancer.Server.Plugins.Leaderboards/Changelog.rst
@@ -9,8 +9,14 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+4.0.2.3
+-------
+Changed
+*******
+- Remove unnecessary logs
+
 4.0.2.2
-----------
+-------
 Changed
 *******
 - Remove debug logs