+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/Changelog.rst
@@ -9,6 +9,14 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+4.1.4.1
+-------
+Changed
+*******
+- Fix handlers (use a request scope in sequential operations)
+- FindGame returns false and message error on failure
+- Add reason on party error (format: party.joinDenied?reason=xxxxxxxx)
+
 4.1.2.1
 -------
 Changed